Waging an unreasonable war on poor invisible trailing whitespace.
Its recommended for everyone who is commiting to turn on the
whitespace checker in Git by doing a 'chmod +x .git/hooks/pre-commit'
from the root of their working repository.  If it ever gets in the way
of your commit and you want to submit the whitespace anyway just
pass the '-n' flag to git commit.

Signed-off-by: Glenn Rempe <glenn@rempe.us>